Soporte para reservoirpy v0.4 (y versionees 0.3.x) Añadido requiremets.txt

+++ icann_process_v2.py
@@ -8,15 +8,22 @@
 # You can also manually download and uncompress the data.
 ###################################################################
 
-# Start the download    
-print('downloading dataset ...')
-import requests
-url = 'https://digibuo.uniovi.es/dspace/bitstream/handle/10651/53461/dataicann.zip?sequence=1&isAllowed=y'
-r = requests.get(url)
-with open('dataicann.zip', 'wb') as outfile:
-    outfile.write(r.content)
-print('download completed')
+from pathlib import Path
 
+file_path = Path('./dataicann.zip')
+
+if not file_path.exists():   
+    # Start the download    
+    print('downloading dataset ...')
+    import requests
+    url = 'https://digibuo.uniovi.es/dspace/bitstream/handle/10651/53461/dataicann.zip?sequence=1&isAllowed=y'
+    r = requests.get(url)
+    with open('dataicann.zip', 'wb') as outfile:
+        outfile.write(r.content)
+    print('download completed')
+else:
+    print('File already exists, skipping download.')
+    
 import zipfile
 with zipfile.ZipFile('./dataicann.zip', 'r') as zip_ref:
     zip_ref.extractall('./')
@@ -27,8 +34,11 @@ with zipfile.ZipFile('./dataicann.zip', 'r') as zip_ref:
 
 
 import reservoirpy as rpy
+from packaging.version import Version
 
-rpy.verbosity(0)  # no need to be too verbose here
+if( Version(rpy.__version__) < Version("0.4")):
+    rpy.verbosity(0)  # no need to be too verbose here
+
 rpy.set_seed(42)  # make everyhting reproducible !
 
 import numpy as np
@@ -81,14 +91,23 @@ ridge = 5.530826061879047e-08
 print('Creating ESN...')
 data = Input()
 reservoir = Reservoir(n_states, lr=Lr, sr=rho, input_scaling=input_scale, rc_connectivity=sparsity, Win=rpy.mat_gen.bernoulli(input_scaling = Win_scale))
-readout = Ridge(ridge=ridge, input_bias = set_bias)
+if(Version(rpy.__version__) < Version("0.4")):
+    readout = Ridge(ridge = ridge, input_bias = set_bias)
+else:
+    readout = Ridge(ridge = ridge, fit_bias = set_bias)
 esn_model =  data >> reservoir >> readout
-print(esn_model.node_names)
+if( Version(rpy.__version__) < Version("0.4")):
+    print(esn_model.node_names)
+else:
+    print(esn_model.nodes)
 
 # Train the ESN with the train data
 print('Training ESN...')
 esn_model = esn_model.fit(X_train, Y_train, warmup=Warmup)
-print(reservoir.is_initialized, readout.is_initialized, readout.fitted)
+if( Version(rpy.__version__) < Version("0.4")):
+    print(reservoir.is_initialized, readout.is_initialized, readout.fitted)
+else:
+    print(reservoir.initialized, readout.initialized)
 
 
 #####################################################################
